Inline Base plugin into core classes, extract them to separate files
