+++ www/index.html
@@ -29,7 +29,7 @@
           <ul>
             {% for item in site.data.core %}
             <li>
-              <a href="https://{{ site.url }}/rdoc/classes/Shrine/{{ item.path }}">
+              <a href="https://{{ site.url }}/rdoc/classes/{{ item.path }}">
                 {{ item.name }}
               </a>
             </li>
